requirements.txt
beautifulsoup4
celery[redis]
emoji
fake-useragent
influxdb
python-telegram-bot
redis
requests
